[tty] Extract tty_handle_default_state() in a separate file
